pyproject.toml
[project]
name = "flax-illuminant-estimation"
version = "0.1.0"
description = "A vision transformer (ViT) model for estimating illuminant color in images"
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "owo93", email = "iztyle30@gmail.com" }]
dependencies = [
  "dotenv>=0.9.9",
  "flax[cuda]>=0.12.3",
  "jax[cuda13]>=0.9.0.1",
  "pillow>=12.1.0",
  "pyyaml>=6.0.3",
  "tqdm>=4.67.3",
  "typer>=0.21.1",
  "typing>=3.10.0.0",
  "wandb>=0.18.0",
]

[project.scripts]
illum = "flax_illuminant_estimation:main"

[dependency-groups]
dev = [
  "ipykernel>=7.2.0",
  "ruff>=0.15.7",
]

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
